examples/api-integration-agno/requirements.txt
# AGNO Framework
agno>=0.1.0

# LLM Provider
openai>=1.50.0

# HTTP Client & Retry Logic
httpx>=0.25.0  # Modern async HTTP client
tenacity>=8.2.3  # Retry logic
requests>=2.31.0  # Fallback HTTP client

# API Development (para sample_api.py)
fastapi>=0.104.0  # Mock API server
uvicorn>=0.24.0  # ASGI server

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0  # Data validation
